requirements.txt
asgiref==3.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
Deprecated==1.2.13
device-detector==5.0.1
Django~=4.0.10
django-extensions==3.2.1
django-filter==23.2
django-oauth-toolkit==2.2.0
djangorestframework==3.14.0
dnspython==2.3.0
idna==3.4
importlib-metadata==6.6.0
jwcrypto==1.4.2
Markdown==3.4.3
mysqlclient==2.1.0
oauthlib==3.2.2
pycparser==2.21
pymongo==3.12.1
pytz==2023.3
PyYAML==6.0
regex==2023.5.5
requests==2.30.0
sqlparse==0.2.4
ua-parser==0.16.1
urllib3==2.0.2
user-agents==2.2.0
wrapt==1.15.0
zipp==3.15.0
drf-yasg==1.21.5
django_mongoengine==0.5.4
django-cors-headers==4.1.0
djangorestframework-simplejwt
PyJWT~=2.8.0